refac: move rust bindings to vendors/halo2
Exposing the same rust types to c++ using `cxx.rs` in different places
is annoying. So this commit collects all the bindings into the same
place where it is really used. This may bring code duplication.
