+++ tachyon/rs/BUILD.bazel
@@ -1,30 +1,12 @@
 load("@crate_index//:defs.bzl", "aliases", "all_crate_deps")
-load("@cxx.rs//tools/bazel:rust_cxx_bridge.bzl", "rust_cxx_bridge")
-load("//bazel:tachyon.bzl", "if_gpu_is_configured")
 load("//bazel:tachyon_rust.bzl", "tachyon_rust_library")
 
 package(default_visibility = ["//visibility:public"])
-
-FEATUERS = if_gpu_is_configured(["gpu"])
 
 tachyon_rust_library(
     name = "tachyon_rs",
     srcs = glob(["src/**/*.rs"]),
     aliases = aliases(),
-    crate_features = FEATUERS,
     proc_macro_deps = all_crate_deps(proc_macro = True),
-    deps = all_crate_deps(normal = True) + [
-        ":bn254_cxx_bridge",
-        "//tachyon/rs/math/elliptic_curves/bn/bn254:msm",
-        "//tachyon/rs/math/elliptic_curves/bn/bn254:msm_gpu",
-    ],
-)
-
-rust_cxx_bridge(
-    name = "bn254_cxx_bridge",
-    src = "src/math/elliptic_curves/bn/bn254/mod.rs",
-    args = if_gpu_is_configured([
-        '--cfg=feature="gpu"',
-    ]),
-    deps = ["//tachyon/rs/math/elliptic_curves/bn/bn254:bn254_hdrs"],
+    deps = all_crate_deps(normal = True),
 )
