--- tachyon/rs/math/elliptic_curves/bn/bn254/BUILD.bazel
@@ -1,34 +1,0 @@
-load("//bazel:tachyon.bzl", "if_gpu_is_configured")
-load("//bazel:tachyon_cc.bzl", "tachyon_cc_library")
-
-package(default_visibility = ["//visibility:public"])
-
-tachyon_cc_library(
-    name = "bn254_hdrs",
-    hdrs = [
-        "msm.h",
-        "msm_gpu.h",
-    ],
-    include_prefix = "tachyon_rs/math/elliptic_curves/bn/bn254",
-    deps = ["@cxx.rs//:core"],
-)
-
-tachyon_cc_library(
-    name = "msm",
-    srcs = ["msm.cc"],
-    deps = [
-        ":bn254_hdrs",
-        "//tachyon/c/math/elliptic_curves/bn/bn254:msm",
-        "//tachyon/rs:bn254_cxx_bridge/include",
-    ],
-)
-
-tachyon_cc_library(
-    name = "msm_gpu",
-    srcs = if_gpu_is_configured(["msm_gpu.cc"]),
-    deps = [
-        ":bn254_hdrs",
-        "//tachyon/c/math/elliptic_curves/bn/bn254:msm_gpu",
-        "//tachyon/rs:bn254_cxx_bridge/include",
-    ],
-)
