+++ vendors/halo2/Cargo.toml
@@ -15,6 +15,7 @@ keywords = ["tachyon", "halo", "proofs", "zkp", "zkSNARKs"]
 publish = false
 
 [dependencies]
+cxx = "1.0"
 halo2_proofs = { git = "https://github.com/kroma-network/halo2.git", rev = "9922fbb" }
 halo2curves = { git = "https://github.com/kroma-network/halo2curves.git", rev = "c0ac193" }
 tachyon_rs = { path = "../../tachyon/rs" }
@@ -23,3 +24,7 @@ tachyon_rs = { path = "../../tachyon/rs" }
 rand_core = { version = "0.6", default-features = false, features = [
     "getrandom",
 ] }
+
+[features]
+default = []
+gpu = []
