+++ vendors/halo2/src/bn254_msm.cc
@@ -1,9 +1,10 @@
-#include "tachyon/rs/math/elliptic_curves/bn/bn254/msm.h"
+#include "vendors/halo2/include/bn254_msm.h"
+
+#include "vendors/halo2/src/bn254.rs.h"
 
 #include "tachyon/c/math/elliptic_curves/bn/bn254/msm.h"
-#include "tachyon/rs/src/math/elliptic_curves/bn/bn254/mod.rs.h"
 
-namespace tachyon::rs::math::bn254 {
+namespace tachyon::halo2_api::bn254 {
 
 rust::Box<G1MSM> create_g1_msm(uint8_t degree) {
   return rust::Box<G1MSM>::from_raw(
@@ -13,18 +14,6 @@ rust::Box<G1MSM> create_g1_msm(uint8_t degree) {
 void destroy_g1_msm(rust::Box<G1MSM> msm) {
   tachyon_bn254_g1_destroy_msm(
       reinterpret_cast<tachyon_bn254_g1_msm_ptr>(msm.into_raw()));
-}
-
-rust::Box<G1JacobianPoint> g1_affine_msm(G1MSM* msm,
-                                         rust::Slice<const G1AffinePoint> bases,
-                                         rust::Slice<const Fr> scalars) {
-  auto ret = tachyon_bn254_g1_affine_msm(
-      reinterpret_cast<tachyon_bn254_g1_msm_ptr>(msm),
-      reinterpret_cast<const tachyon_bn254_g1_affine*>(bases.data()),
-      reinterpret_cast<const tachyon_bn254_fr*>(scalars.data()),
-      scalars.length());
-  return rust::Box<G1JacobianPoint>::from_raw(
-      reinterpret_cast<G1JacobianPoint*>(ret));
 }
 
 rust::Box<G1JacobianPoint> g1_point2_msm(G1MSM* msm,
@@ -39,4 +28,4 @@ rust::Box<G1JacobianPoint> g1_point2_msm(G1MSM* msm,
       reinterpret_cast<G1JacobianPoint*>(ret));
 }
 
-}  // namespace tachyon::rs::math::bn254
+}  // namespace tachyon::halo2_api::bn254
