+++ vendors/halo2/src/lib.rs
@@ -1,1 +1,2 @@
+mod bn254;
 mod msm;
